refacto: separation of tests according to new field
