+++ tests/Controller/TaskControllerTest.php
@@ -16,12 +16,11 @@ class TaskControllerTest extends WebTestCase
     private TaskRepository $taskRepository;
     private UserRepository $userRepository;
     private User $user;
-    private User $admin;
-    private User $userWithoutRole;
     private Task $taskUser1;
     private Task $taskUser2;
     private Task $task;
     private Task $anonymousTask;
+
 
     protected function setUp(): void
     {
@@ -44,43 +43,6 @@ class TaskControllerTest extends WebTestCase
  
         $user = $this->userRepository->findOneByEmail('user1@todolist.fr');
         $this->user = $user;
-        $user = $this->userRepository->findOneByEmail('user3@todolist.fr');
-        $this->userWithoutRole = $user;
-        $admin = $this->userRepository->findOneByEmail('admin@todolist.fr');
-        $this->admin = $admin;
-    }
-
-    public function testCreateTaskWithUnauthorizedAccess(): void
-    {
-        $this->client->followRedirects();
-
-        $this->client->request('GET', '/tasks/create');
-        $response = $this->client->getResponse();
- 
-        $this->assertEquals(Response::HTTP_OK, $response->getStatusCode());
-        $this->assertSelectorTextContains('button', 'Se connecter');
-    }
-
-    public function testCreateTaskWithAuthorizedAccess(): void
-    {
-        $this->client->loginUser($this->user, 'secured_area');
-
-        $this->client->request('GET', '/tasks/create');
-        $response = $this->client->getResponse();
- 
-        $this->assertEquals(Response::HTTP_OK, $response->getStatusCode());
-        $this->assertSelectorTextContains('button', 'Ajouter');
-    }
-
-    public function testCreateTasktWithDataWithUserWithoutRole()
-    {   
-        $this->client->followRedirects();
-        
-        $this->client->loginUser($this->userWithoutRole, 'secured_area'); 
-        $this->client->request('GET', '/tasks/create');
-        $response = $this->client->getResponse();
- 
-        $this->assertEquals(Response::HTTP_FORBIDDEN, $response->getStatusCode());
     }
 
     public function testCreateTasktWithData()
@@ -106,58 +68,6 @@ class TaskControllerTest extends WebTestCase
         $this->assertSelectorTextContains('h1', 'Liste des tâches');
     }
 
-    public function testEditTaskWithUnauthorizedAccess(): void
-    {
-        $this->client->followRedirects();
-
-        $this->client->request('GET', '/tasks/'.$this->task->getId().'/edit');
-        $response = $this->client->getResponse();
- 
-        $this->assertEquals(Response::HTTP_OK, $response->getStatusCode());
-        $this->assertSelectorTextContains('button', 'Se connecter');
-    }
-
-    public function testEditTasktWitUnauthorizedUser()
-    {
-        $this->client->loginUser($this->user, 'secured_area');
-
-        $this->client->request('GET', '/tasks/'.$this->taskUser2->getId().'/edit');
-        $response = $this->client->getResponse();
- 
-        $this->assertEquals(Response::HTTP_FORBIDDEN, $response->getStatusCode());
-    }
-
-    public function testEditTaskWithOutOwnerAndWithRoleUser()
-    {
-        $this->client->loginUser($this->user, 'secured_area');
-
-        $this->client->request('GET', '/tasks/'.$this->anonymousTask->getId().'/edit');
-        $response = $this->client->getResponse();
- 
-        $this->assertEquals(Response::HTTP_FORBIDDEN, $response->getStatusCode());
-    }
-
-    public function testEditTaskWithOutOwnerAndWithRoleAdmin()
-    {
-        $this->client->loginUser($this->admin, 'secured_area');
-
-        $this->client->request('GET', '/tasks/'.$this->anonymousTask->getId().'/edit');
-        $response = $this->client->getResponse();
- 
-        $this->assertEquals(Response::HTTP_OK, $response->getStatusCode());
-        $this->assertSelectorTextContains('button', 'Modifier');
-    }
-
-    public function testEditTasktWithAuthorizedUser()
-    {
-        $this->client->loginUser($this->user, 'secured_area');
-        $this->client->request('GET', '/tasks/'.$this->taskUser1->getId().'/edit');
-        $response = $this->client->getResponse();
-
-        $this->assertEquals(Response::HTTP_OK, $response->getStatusCode());
-        $this->assertSelectorTextContains('button', 'Modifier');
-    }
-
     public function testEditTasktWithNewData()
     {   
         $this->client->followRedirects();
@@ -177,65 +87,6 @@ class TaskControllerTest extends WebTestCase
         $crawler = $this->client->submit($form);
 
         $this->assertResponseIsSuccessful();
-        $this->assertSelectorTextContains('h1', 'Liste des tâches');
-    }
-
-    public function testTaskDeleteWithUnauthorizedAccess()
-    {        $this->client->followRedirects();
-        $this->client->request('GET', '/tasks/'.$this->task->getId().'/delete');
-        $response = $this->client->getResponse();
-
-        $this->assertEquals(Response::HTTP_OK, $response->getStatusCode());
-        $this->assertSelectorTextContains('button', 'Se connecter');
-    }
-
-    public function testTaskDeleteWithUnauthorizedUser()
-    {
-        $this->client->loginUser($this->user, 'secured_area');
-
-        $this->client->request('GET', '/tasks/'.$this->taskUser2->getId().'/delete');
-        $response = $this->client->getResponse();
-
-        $this->assertEquals(Response::HTTP_FORBIDDEN, $response->getStatusCode());
-    }
-
-    public function testAnonymousTaskDeleteWithUnauthorizedUser()
-    {
-        $this->client->loginUser($this->user, 'secured_area');
-
-        $this->client->request('GET', '/tasks/'.$this->anonymousTask->getId().'/delete');
-        $response = $this->client->getResponse();
-
-        $this->assertEquals(Response::HTTP_FORBIDDEN, $response->getStatusCode());
-    }
-
-    public function testTaskDeleteWithAuthorizedAccess()
-    {
-        $this->client->followRedirects();
-        $this->client->loginUser($this->user, 'secured_area'); 
-        $taskToDelete = $this->taskRepository->findOneByTitle('Titre tache utilisateur 1 à supprimer');
-
-        $this->client->request('GET', '/tasks/'.$taskToDelete->getId().'/delete');
-        $response = $this->client->getResponse();
-
-        $this->assertResponseIsSuccessful();
-        $this->assertEquals(Response::HTTP_OK, $response->getStatusCode());
-
-        $this->assertSelectorTextContains('h1', 'Liste des tâches');
-    }
-
-    public function testTaskDeleteWithRoleAdminInOtherOwner()
-    {
-        $this->client->followRedirects();
-        $this->client->loginUser($this->admin, 'secured_area');
-        $taskToDelete = $this->taskRepository->findOneByTitle('Titre tache utilisateur 2 à supprimer');
-
-        $this->client->request('GET', '/tasks/'.$taskToDelete->getId().'/delete');
-        $response = $this->client->getResponse();
-
-        $this->assertResponseIsSuccessful();
-        $this->assertEquals(Response::HTTP_OK, $response->getStatusCode());
-
         $this->assertSelectorTextContains('h1', 'Liste des tâches');
     }
 
@@ -266,18 +117,6 @@ class TaskControllerTest extends WebTestCase
     {
         $this->client->followRedirects();
         $this->client->loginUser($this->user, 'secured_area');
-
-        $this->client->request('GET', '/tasks/'.$this->anonymousTask->getId().'/toggle');
-        $response = $this->client->getResponse();
- 
-        $this->assertEquals(Response::HTTP_OK, $response->getStatusCode());
-        $this->assertSelectorTextContains('h1', 'Liste des tâches');
-    }
-
-    public function testToggleTaskWithOutOwnerAndWithRoleAdmin()
-    {
-        $this->client->followRedirects();
-        $this->client->loginUser($this->admin, 'secured_area');
 
         $this->client->request('GET', '/tasks/'.$this->anonymousTask->getId().'/toggle');
         $response = $this->client->getResponse();
